Moving code review save log config to specific handler
